-adaptation of DVD module to navigation slider. The seek has to be done on a multiple of 2048, and shopuld take care of the offset to the beginning of the stream. It works with most of my DVDs by fails to detect size of some streams.
-change of the order of call to ifo and css functions in input_dvd.c,
because we can't read ifo files before authentification on some DVDs.

-changed location of some css tables.

-we read ifo date and decrypt css keys for all titles now. It takes quite a
long time to initialize with some DVDs (if you have 42 titles for
instance). The key decryption might give a false value if the title is
less than 10 Mb ; I'm working on that :)

+++ plugins/dvd/dvd_ifo.h
@@ -2,7 +2,7 @@
  * dvd_ifo.h: Structures for ifo parsing
  *****************************************************************************
  * Copyright (C) 1999-2001 VideoLAN
- * $Id: dvd_ifo.h,v 1.3 2001/02/09 03:51:42 stef Exp $
+ * $Id: dvd_ifo.h,v 1.4 2001/02/13 10:08:51 stef Exp $
  *
  * Author: Stéphane Borel <stef@via.ecp.fr>
  *
@@ -451,7 +451,7 @@ typedef struct vts_tmap_ti_s
  */
 typedef struct vts_s
 {
-    u32             i_pos;
+    off_t           i_pos;
     vtsi_mat_t      mat;
     /* Part Of Title Search Pointer Table Info */
     vts_ptt_srpt_t  ptt_srpt;
